refactor: migrate from FastMCP to official @modelcontextprotocol/sdk
This fixes tool registration issues with Claude Desktop by using the same
MCP library that working servers (like gtoolsmcp) use.

- Replace fastmcp dependency with @modelcontextprotocol/sdk
- Rewrite server.ts to use official SDK patterns
- Consolidate tools and prompts into server.ts
- Remove separate tool and prompt module files
- Update biome config to ignore dist/
- Bump version to 1.2.0
