+++ biome.json
@@ -7,7 +7,7 @@
 	},
 	"files": {
 		"ignoreUnknown": false,
-		"ignore": []
+		"ignore": ["dist/", "node_modules/"]
 	},
 	"formatter": {
 		"enabled": true,
