--- src/prompts/index.ts
@@ -1,4 +1,0 @@
-export { dailySummaryPrompt } from "./dailySummary.js";
-export { researchNotePrompt } from "./researchNote.js";
-export { meetingNotesPrompt } from "./meetingNotes.js";
-export { jobApplicationPrompt } from "./jobApplication.js";
