--- src/tools/index.ts
@@ -1,6 +1,0 @@
-export { listSpacesTool } from "./listSpaces.js";
-export { getSpaceInfoTool } from "./getSpaceInfo.js";
-export { searchTool } from "./search.js";
-export { saveWeblinkTool } from "./saveWeblink.js";
-export { saveToDailyNoteTool } from "./saveToDailyNote.js";
-export { readObjectContentTool } from "./readObjectContent.js";
